Create compat header file
